+++ src/components/dialogs/api-key-dialog.tsx
@@ -8,10 +8,10 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
-} from "./ui/dialog";
-import { Label } from "./ui/label";
-import { Input } from "./ui/input";
-import { Button } from "./ui/button";
+} from "@/ui/dialog";
+import { Label } from "@/ui/label";
+import { Input } from "@/ui/input";
+import { Button } from "@/ui/button";
 import React from "react";
 import { useStore } from "@/hooks/use-store";
