+++ src/components/selectors/safe-mode-selector.tsx
