+++ src/components/selectors/temperature-selector.tsx
